BCDA-2941 EOY Changes Ending: Updating static site guide and updates for the runouts feature (#115)
* Removing all EOY alerts

There were 4 instances of the same alert, telling users not to use anything except the runouts temporarily, these have been removed.

* Add update to Home Page and Updates for runouts

Added a quick update to inform users that BCDA has returned to normal operation after receiving the January attribution file.

* Updates language change

Edited the language within the updates to better reflect past posts

* Update the build.html page to rearrange sections

Change the format of the Build page to lower the emphasis on the EOY section. Moved EOY section to the Requesting Filtered Data chapter.

* Update /build page, EOY section language

Removed all mention of 400 Bad Requests when runouts was not included. Deleted Step 3 and 4 and replaced with references to Above links.

* Update Requesting Filtered Data Intro to include use cases for runouts.

Add the runouts use cases to the first paragraph. Added runouts as list item in the intro. Slightly changed other language to accomodate more than _since.

* Slight updates to language

Changed language in the headers of the EOY sections to match others.

* Anchor links and clarification of runout behavior

* Adding links to build page

* Changing file names and references

Changed EOY to runout in all file names and references

* Clean up files from old site

Deleted the prod folder and the Add to Guide folder

Co-authored-by: Dave Sukharan <davesukharan@davesukadhocmbp.lan>
Co-authored-by: Elissa Frankle Olinsky <elissa@adhocteam.us>