BCDA-7991: Address sonarqube security issues (#195)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7991

## 🛠 Changes

Updated Dockerfile + problematic html file.

## ℹ️ Context for reviewers

SonarQube had 8 "findings" related to the static site application: 2
related to the Dockerfile in the project, and 6 related to external link
outs. 7 of those findings will be resolved with this PR, and one can
remain ignored (the copying all files in a directory in the Dockerfile)
as it's mitigated by addressing the second Dockerfile finding (ensuring
that the docker container runs in user mode)

## ✅ Acceptance Validation

Addressed/mitigated each security issue by ensuring the Dockerfile
specifies the use of a non-root user and adding rel="noopener" tags.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.