PLT-469: Updated pull request template and added CONTRIBUTING.md (#203)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-469

## 🛠 Changes

Added the updated pull request template and contribution guidelines from
[ab2d-bcda-dpc-platform
#87](https://github.com/CMSgov/ab2d-bcda-dpc-platform/pull/87)

## ℹ️ Context for reviewers

Jason is rolling off of the project, so we're updating the pull requests
templates to reflect our new ISSO @SJWalter11. We're also taking this
opportunity to introduce a new contribution guidelines file, as to not
clog up the average PR with extra steps.

## ✅ Acceptance Validation

The pull request template should be reflected on the creation of new
PRs.