Update partially adjudicated claims data sets - enrollee count (#208)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8288

## 🛠 Changes

Small REACH ACO - 100 → 110
Large REACH ACO - 10,000 → 11,000

## ℹ️ Context

Website needs to accurately reflect the number of enrollees in the
synthetic data sets after the changes in
[BCDA-8125](https://jira.cms.gov/browse/BCDA-8125).

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->