BCDA-8390-adjust-file-download-copy (#219)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8390

## 🛠 Changes

Adjusting the file download curl command to specifically download a file
as output instead of return bits from the request as well as adjust the
copy to acknowledge that the file downloaded is a gzip and will need to
be extracted to be used.

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Site build and review locally.