Get a bearer token page  (#230)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8856

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->

## ℹ️ Context

Line 46 surfaces a gitleaks error. [Read the gitleaks
docs](https://github.com/gitleaks/gitleaks?tab=readme-ov-file)

I attempted to fix this error with both `#gitleaks:allow` as an in-line
comment and with a '.gitleaksignore' file, but since this is a code
snippet captured as '{% raw %}' neither of these workarounds were
successful.

Therefore, I fixed this my removing two characters in line 46, which
prevented gitleaks from recognizing it as a leak. Since this snippet is
purely for example purposes, this shouldn't be an issue

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

---------

Co-authored-by: phamjennifer <phammjennifer@gmail.com>