BCDA-9396: create docker configuration for static site (#280)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9396

## 🛠 Changes

- reorganized repo (added a `src` folder to facilitate docker assembly)
- added docker configuration (dockerfile, docker-compose and
dockerignore)
- updated readme, workflow(s), and references to paths appropriately

## ℹ️ Context for reviewers

(Background context, more in-depth details of the implementation, and
anything else you'd like to call out or ask reviewers.)

This PR introduces a docker image for building/running the BCDA static
site. The site is assembled in a docker build very similarly to how it's
built for production.

## ✅ Acceptance Validation

(How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable.)

1. run `docker compose up --watch` to run the site
2. make changes to the site source code and observe that the changes are
reflected in the browser

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.

---------

Co-authored-by: Jonathan Hutchison <jonathanhutchison@navapbc.com>