Update deployment paths in workflows (#281)
## 🎫 Ticket


## 🛠 Changes

- updated the deploy workflow (and other workflows) to associate PR
deployments with a subpath under `/feature/{PR#}`
- updated the deploy workflow (and other workflows) to associate
deployments from main with the staging root path `/`

## ℹ️ Context for reviewers

Presently, when changes are merged to main, they are deployed to the url
https://stage.bcda.cms.gov/main/. This paradigm adds some difficulty to
the prod promotion process:
1. the base url of the main deployment in staging (`/main`) is different
from that of prod (`/`), which means that for every promotion, and
temporary change must be introduced tot he staging workflow to remove
the base url
2. the exact build promoted to prod is _not_ identical to the build
tested in staging, since the staging build must be reproduced in the
above point, and that build has broken links due to the mismatch between
the built base url and the deployed directory structure

This update ensures that merges to main are instead deployed to
https://stage.bcda.cms.gov/, which keeps the base staging url up to date
with changes that will go to prod, and makes deployment to prod easier
since devs won't have to constantly make and remove similar changes to
the deploy workflow. It also ensures that what is tested in staging is
exactly what is deployed to prod.
Additionally, PR changes are deployed to `/feature/{PR#}` instead of
`/{PR#}`, which makes it simple to exclude them from the prod promotion.
When a staging deployment is promoted to prod, the immediate
subdirectory `/feature` is ignored, enabling users to safely deploy
feature branches to `/feature` while keeping it separate from anything
that might be deployed to prod.

## ✅ Acceptance Validation

(How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable.)

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.