BCDA-9416 Add note on adjudicated claims availability (#285)
## 🎫 Ticket

https://jira.cms.gov/browse/bcda-9416

## 🛠 Changes

Add short note letting customers know when to expect updates to
adjudicated claims data to be available via BCDA

## ℹ️ Context

Previously we had this information on the website, but it was removed in
the website redesign

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

---------

Co-authored-by: Jonathan Hutchison <jonathanhutchison@navapbc.com>